refactor: align dashboard feature structure
