+++ src/features/dashboard/index.ts
@@ -1,5 +1,5 @@
-export { default as DashboardContainer } from './container/DashboardContainer';
-export { default as DashboardPage } from './pages/DashboardPage';
+export { default as DashboardContainer } from './containers/DashboardContainer';
+export * from './routes/dashboard.routes';
 export * from './components';
 export * from './hooks';
 export * from './services';
