--- src/features/dashboard/pages/DashboardPage.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import DashboardContainer from '../container/DashboardContainer';
-
-const DashboardPage: React.FC = () => {
-  return <DashboardContainer />;
-};
-
-export default DashboardPage;
