+++ src/features/dashboard/routes/dashboard.routes.ts
@@ -1,0 +1,3 @@
+export const DASHBOARD_ROUTES = {
+  ROOT: '/dashboard',
+} as const;
